requirements.txt
Django==1.8
South==1.0.2
mysql-python==1.2.5
django-extensions==1.5.5
requests==2.7.0
djangorestframework==3.1.2
gunicorn==19.4.5
